Fix fs_nmea.SPACETIME
Renamed nmea to fs_nmea
Added new dictionaries to __init__.py
GPS dictionary is master of all dictionaries
