+++ backend/scripts/dicts/GPS.py
@@ -1,0 +1,15 @@
+'''
+Farmstar blank dictionary of dictionaries
+'''
+
+GPS = {'STATUS':'',
+       'SPACETIME':'',
+       'GGA':'',
+       'GSA':'',
+       'GSV':'',
+       'RMC':'',
+       'HDG':'',
+       'HDT':'',
+       }
+       
+
