+++ backend/scripts/dicts/STATUS.py
@@ -1,0 +1,13 @@
+'''
+Farmstar blank dictionary for status messages\
+Append as required
+'''
+
+STATUS = {'status':{},
+          'string':{},
+          'check':{},
+          'cpu':{},
+          'ram':{},
+          }
+       
+
